+++ src/main/kotlin/icu/windea/pls/ep/inspections/ParadoxIncorrectExpressionCheckers.Basic.kt
@@ -7,13 +7,9 @@ import icu.windea.pls.base.annotations.WithGameType
 import icu.windea.pls.config.CwtDataTypeSets
 import icu.windea.pls.config.CwtDataTypes
 import icu.windea.pls.config.config.CwtMemberConfig
-import icu.windea.pls.config.config.CwtPropertyConfig
-import icu.windea.pls.config.config.CwtValueConfig
-import icu.windea.pls.config.config.aliasConfig
+import icu.windea.pls.config.config.expandConfigExpression
 import icu.windea.pls.config.configExpression.floatRange
 import icu.windea.pls.config.configExpression.intRange
-import icu.windea.pls.core.castOrNull
-import icu.windea.pls.core.collections.orNull
 import icu.windea.pls.core.isExactDigit
 import icu.windea.pls.core.unquote
 import icu.windea.pls.csv.psi.ParadoxCsvExpressionElement
@@ -23,31 +19,28 @@ import icu.windea.pls.lang.psi.ParadoxExpressionElement
 import icu.windea.pls.lang.psi.floatValue
 import icu.windea.pls.lang.psi.intValue
 import icu.windea.pls.lang.psi.resolved
-import icu.windea.pls.lang.psi.stringValue
 import icu.windea.pls.lang.resolve.complexExpression.ParadoxScopeFieldExpression
 import icu.windea.pls.lang.search.ParadoxDefinitionSearch
-import icu.windea.pls.lang.select.selectScope
 import icu.windea.pls.lang.util.ParadoxScopeManager
 import icu.windea.pls.lang.util.evaluators.ParadoxComplexExpressionEvaluator
 import icu.windea.pls.model.ParadoxGameType
 import icu.windea.pls.model.scope.ParadoxScopeContext
-import icu.windea.pls.model.type.CwtExpressionType
 import icu.windea.pls.script.psi.ParadoxScriptColor
 import icu.windea.pls.script.psi.ParadoxScriptExpressionElement
 import icu.windea.pls.script.psi.ParadoxScriptMember
-import icu.windea.pls.script.psi.ParadoxScriptScriptedVariableReference
-import icu.windea.pls.script.psi.ParadoxScriptString
 import icu.windea.pls.script.psi.ParadoxScriptStringExpressionElement
 
 /**
  * 检查整数数值是否在指定的区间内。
  */
-class ParadoxRangedIntChecker : ParadoxIncorrectExpressionChecker {
+class ParadoxRangedIntFieldChecker : ParadoxIncorrectExpressionChecker {
     override fun check(element: ParadoxExpressionElement, config: CwtMemberConfig<*>, context: ParadoxExpressionInspectionContext): Boolean {
         if (element !is ParadoxScriptExpressionElement && element !is ParadoxCsvExpressionElement) return true
 
-        val configExpression = config.configExpression
-        if (configExpression.type !in CwtDataTypeSets.IntField) return true // for int only
+        // for int field only (after expansion)
+        val configExpression = config.expandConfigExpression().find { it.type in CwtDataTypeSets.IntField }
+        if (configExpression == null) return true
+
         val intRange = configExpression.intRange ?: return true
         val intValue = when {
             element is ParadoxScriptExpressionElement -> element.resolved()?.intValue()
@@ -63,12 +56,14 @@ class ParadoxRangedIntChecker : ParadoxIncorrectExpressionChecker {
 /**
  * 检查浮点数数值是否在指定的区间内。
  */
-class ParadoxRangedFloatChecker : ParadoxIncorrectExpressionChecker {
+class ParadoxRangedFloatFieldChecker : ParadoxIncorrectExpressionChecker {
     override fun check(element: ParadoxExpressionElement, config: CwtMemberConfig<*>, context: ParadoxExpressionInspectionContext): Boolean {
         if (element !is ParadoxScriptExpressionElement && element !is ParadoxCsvExpressionElement) return true
 
-        val configExpression = config.configExpression
-        if (configExpression.type !in CwtDataTypeSets.FloatField) return true // for int and float
+        // for float field only (after expansion)
+        val configExpression = config.expandConfigExpression().find { it.type in CwtDataTypeSets.FloatField }
+        if (configExpression == null) return true
+
         val floatRange = configExpression.floatRange ?: return true
         val floatValue = when {
             element is ParadoxScriptExpressionElement -> element.resolved()?.floatValue()
@@ -88,8 +83,10 @@ class ParadoxColorFieldChecker : ParadoxIncorrectExpressionChecker {
     override fun check(element: ParadoxExpressionElement, config: CwtMemberConfig<*>, context: ParadoxExpressionInspectionContext): Boolean {
         if (element !is ParadoxScriptColor) return true
 
-        val configExpression = config.configExpression
-        if (configExpression.type != CwtDataTypes.ColorField) return true
+        // for color field only (after expansion)
+        val configExpression = config.expandConfigExpression().find { it.type == CwtDataTypes.ColorField }
+        if (configExpression == null) return true
+
         val expectedColorType = configExpression.value ?: return true
         val colorType = element.colorType
         if (colorType == expectedColorType) return true
@@ -106,16 +103,16 @@ class ParadoxScopeBasedScopeFieldExpressionChecker : ParadoxIncorrectExpressionC
     override fun check(element: ParadoxExpressionElement, config: CwtMemberConfig<*>, context: ParadoxExpressionInspectionContext): Boolean {
         if (element !is ParadoxScriptStringExpressionElement) return true
 
-        val configExpression = config.configExpression
-        if (configExpression.type != CwtDataTypes.Scope) return true
+        val configExpression = config.expandConfigExpression().find { it.type == CwtDataTypes.Scope }
+        if (configExpression == null) return true
+
         val expectedScope = configExpression.value ?: return true
         val value = element.value
-        val configGroup = config.configGroup
-        val scopeFieldExpression = ParadoxScopeFieldExpression.resolve(value, null, configGroup) ?: return true
+        val scopeFieldExpression = ParadoxScopeFieldExpression.resolve(value, null, context.configGroup) ?: return true
         val memberElement = element.parentOfType<ParadoxScriptMember>(withSelf = true) ?: return true
         val parentScopeContext = ParadoxScopeManager.getScopeContext(memberElement) ?: ParadoxScopeContext.resolveAny()
         val scopeContext = ParadoxScopeManager.getScopeContext(element, scopeFieldExpression, parentScopeContext)
-        if (ParadoxScopeManager.matchesScope(scopeContext, expectedScope, configGroup)) return true
+        if (ParadoxScopeManager.matchesScope(scopeContext, expectedScope, context.configGroup)) return true
         val description = ChronicleEpBundle.message("incorrectExpression.scope.desc", expectedScope, scopeContext.scope.id)
         context.holder.registerProblem(element, description)
         return false
@@ -129,8 +126,9 @@ class ParadoxScopeGroupBasedScopeFieldExpressionChecker : ParadoxIncorrectExpres
     override fun check(element: ParadoxExpressionElement, config: CwtMemberConfig<*>, context: ParadoxExpressionInspectionContext): Boolean {
         if (element !is ParadoxScriptStringExpressionElement) return true
 
-        val configExpression = config.configExpression
-        if (configExpression.type != CwtDataTypes.ScopeGroup) return true
+        val configExpression = config.expandConfigExpression().find { it.type == CwtDataTypes.ScopeGroup }
+        if (configExpression == null) return true
+
         val expectedScopeGroup = configExpression.value ?: return true
         val value = element.value
         val configGroup = config.configGroup
@@ -152,8 +150,9 @@ class ParadoxIntValueFieldChecker : ParadoxIncorrectExpressionChecker {
     override fun check(element: ParadoxExpressionElement, config: CwtMemberConfig<*>, context: ParadoxExpressionInspectionContext): Boolean {
         if (element !is ParadoxScriptExpressionElement) return true
 
-        val configExpression = config.configExpression
-        if (configExpression.type != CwtDataTypes.IntValueField) return true
+        // for int value field only (after expansion)
+        val configExpression = config.expandConfigExpression().find { it.type == CwtDataTypes.IntValueField }
+        if (configExpression == null) return true
 
         val evaluated = ParadoxComplexExpressionEvaluator().evaluate(element) ?: return true
         val intValue = evaluated.intValue()
@@ -179,8 +178,9 @@ class ParadoxFloatValueFieldChecker : ParadoxIncorrectExpressionChecker {
     override fun check(element: ParadoxExpressionElement, config: CwtMemberConfig<*>, context: ParadoxExpressionInspectionContext): Boolean {
         if (element !is ParadoxScriptExpressionElement) return true
 
-        val configExpression = config.configExpression
-        if (configExpression.type != CwtDataTypes.ValueField) return true
+        // for value field only (after expansion)
+        val configExpression = config.expandConfigExpression().find { it.type == CwtDataTypes.ValueField }
+        if (configExpression == null) return true
 
         val evaluated = ParadoxComplexExpressionEvaluator().evaluate(element) ?: return true
         val floatValue = evaluated.floatValue()
@@ -208,8 +208,9 @@ class StellarisTechnologyWithLevelChecker : ParadoxIncorrectExpressionChecker {
     override fun check(element: ParadoxExpressionElement, config: CwtMemberConfig<*>, context: ParadoxExpressionInspectionContext): Boolean {
         if (element !is ParadoxScriptStringExpressionElement) return true
 
-        val configExpression = config.configExpression
-        if (configExpression.type != CwtDataTypes.TechnologyWithLevel) return true
+        val configExpression = config.expandConfigExpression().find { it.type == CwtDataTypes.TechnologyWithLevel }
+        if (configExpression == null) return true
+
         val (technologyName, technologyLevel) = element.value.split('@', limit = 2).takeIf { it.size == 2 } ?: return true
         val project = config.configGroup.project
         val text = element.text
@@ -228,84 +229,3 @@ class StellarisTechnologyWithLevelChecker : ParadoxIncorrectExpressionChecker {
     }
 }
 
-/**
- * 检查特定语句中指定的触发器（trigger）是否是简单触发器（simple trigger）。
- *
- * 适用于：
- * - `switch = {...}`
- * - `inverted_switch = {...}`
- */
-class ParadoxTriggerInSwitchStatementsChecker : ParadoxIncorrectExpressionChecker {
-    object Constants {
-        val triggerKeys = setOf("trigger", "on_trigger")
-        val contextNames = setOf("switch", "inverted_switch")
-    }
-
-    override fun check(element: ParadoxExpressionElement, config: CwtMemberConfig<*>, context: ParadoxExpressionInspectionContext): Boolean {
-        if (element !is ParadoxScriptExpressionElement) return true
-
-        if (element !is ParadoxScriptString && element !is ParadoxScriptScriptedVariableReference) return true
-        if (config !is CwtValueConfig || config.value != "alias_keys_field[trigger]") return true
-
-        val propertyConfig = config.propertyConfig ?: return true
-        if (propertyConfig.key !in Constants.triggerKeys) return true
-        val aliasConfig = propertyConfig.parentConfig?.castOrNull<CwtPropertyConfig>()?.aliasConfig ?: return true
-        if (aliasConfig.subName !in Constants.contextNames) return true
-
-        val triggerName = element.stringValue() ?: return true
-        val configGroup = config.configGroup
-        val resultTriggerConfigs = configGroup.aliasGroups.get("trigger")?.get(triggerName)?.orNull() ?: return true
-
-        if (resultTriggerConfigs.none { it.config.valueType != CwtExpressionType.Block }) {
-            context.holder.registerProblem(element, ChronicleEpBundle.message("incorrectExpression.simpleTrigger.desc", element.expression))
-        }
-        return true
-    }
-}
-
-/**
- * 检查特定语句中指定的触发器（trigger）是否是简单触发器（simple trigger）（如果不带参数），或者复杂触发器（complex trigger）（如果带参数）。
- *
- * 适用于：
- * - `complex_trigger_modifier = {...}`
- * - `export_trigger_value_to_variable = {...}`
- */
-class ParadoxTriggerInWithParametersStatementsChecker : ParadoxIncorrectExpressionChecker {
-    object Constants {
-        const val triggerKey = "trigger"
-        val contextNames = setOf("complex_trigger_modifier", "export_trigger_value_to_variable")
-    }
-
-    override fun check(element: ParadoxExpressionElement, config: CwtMemberConfig<*>, context: ParadoxExpressionInspectionContext): Boolean {
-        // `__` - caret position
-        // `<container> = { <trigger_field> = __<trigger> parameters = { ... } }`
-        // -> `<container> = { <trigger_field> = <trigger> __parameters = { ... } }`
-
-        if (element !is ParadoxScriptExpressionElement) return true
-
-        if (element !is ParadoxScriptString && element !is ParadoxScriptScriptedVariableReference) return true
-        if (config !is CwtValueConfig || config.value != "alias_keys_field[trigger]") return true
-
-        val propertyConfig = config.propertyConfig ?: return true
-        if (propertyConfig.key != Constants.triggerKey) return true
-        val aliasConfig = propertyConfig.parentConfig?.castOrNull<CwtPropertyConfig>()?.aliasConfig ?: return true
-        if (aliasConfig.subName !in Constants.contextNames) return true
-
-        val triggerName = element.stringValue() ?: return true
-        val configGroup = config.configGroup
-        val resultTriggerConfigs = configGroup.aliasGroups.get("trigger")?.get(triggerName)?.orNull() ?: return true
-
-        val hasParameters = selectScope { element.queryParentBy("*/*").asProperty().queryBy("parameters").asProperty().any() }
-        if (hasParameters) {
-            if (resultTriggerConfigs.none { it.config.valueType == CwtExpressionType.Block }) {
-                context.holder.registerProblem(element, ChronicleEpBundle.message("incorrectExpression.complexTrigger.desc", element.expression))
-            }
-        } else {
-            // can also be complex trigger here, for some parameters can be ignored (like `count = xxx`)
-            // if (resultTriggerConfigs.none { !it.config.isBlock }) {
-            //    context.holder.registerProblem(element, ChronicleEpBundle.message("incorrectExpression.checker.expect.simpleTrigger", element.expression.orEmpty()))
-            // }
-        }
-        return true
-    }
-}